Added entities and Product interface

+++ inventory-manager-service/src/main/java/dev/ak/inventoryManager/entity/Batch.java
@@ -15,6 +15,6 @@ public abstract class Batch {
     private Long id;
     private String nameOfBatch;
     private LocalDateTime  dateOfReceipt;
-    private int quantity;
+    private double quantity;
     private BigDecimal costPrice;
 }
